+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [2.0.0] - 2020-07-18
+
 ### Added
 
 - Added config for React apps and libraries
@@ -26,5 +28,6 @@ and this project adheres to
 
 Initial release.
 
-[unreleased]: https://github.com/dxw/eslint-config/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/dxw/eslint-config/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/dxw/eslint-config/compare/v1.0.0...v2.0.0
 [1.0.0]: https://github.com/dxw/eslint-config/releases/tag/v1.0.0